add: cmd/dig: shared handler
